requirements.txt
absl-py==1.4.0
albumentations==1.3.0
appdirs==1.4.4
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
coloredlogs==15.0.1
colorlog==6.7.0
cycler==0.11.0
dacite==1.7.0
decorator==5.1.1
dparse==0.6.2
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.25.0
importlib-metadata==6.0.0
ipython==7.34.0
jedi==0.18.2
joblib==1.2.0
kiwisolver==1.4.4
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mpmath==1.2.1
networkx==2.6.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
onnx==1.13.0
onnxruntime-gpu==1.13.1
opencv-python==4.7.0.68
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycuda==2022.1
Pygments==2.14.0
pymediainfo==6.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytools==2022.1.12
pytz==2022.7.1
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
safety==2.3.5
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
sympy==1.10.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2021.11.2
toml==0.10.2
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.2
xmltodict==0.13.0
zipp==3.11.0
